Change function to check for next line
Changed the getContinuousUserInfo function to now check that there is a new line.
Previously, it tried to get the new line without checking if it existed first.
This fixed a problem with my not so great testing method of piping a text file
into the program where it would crash when trying to get entry info.